requirements.txt
pandas
numpy
python-dotenv
pymysql
imblearn
fastapi
pydantic